backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
Pillow==11.3.0
PyJWT==2.10.1
